Make logout link return to the login screen
Originally, I tried to wire up the logout link to use the React
`onClick` event to return to the login screen. This did not work because
the dlmenu plugin prevents click events from being propagated.

See this issue for more details:
https://github.com/codrops/ResponsiveMultiLevelMenu/issues/29

I fixed this bug in our version of dlmenu but decided to use another
approach altogether. Instead of using React 'onClick', I decided to add
a `[data-url]` attribute to the logout link. However, it appears that
this does not work because React handles all click events on the `body`
DOM element. The workaround I discovered was to update our code to
listen for `[data-url]` link clicks on the `#container` element instead of
`body`.