Render modals and provide S3 upload functionality.
We render our modals into the document before any other elements. This is
considered the safest way to work with Bootstrap modals. The modal HTML for
the upload form is converted to a React element. A file upload button is added
to it that is integrated with the S3 functionality.