Get AWS settings via boto.
It will be better if we can allow boto to do the signing for us. Boto can
normalize some aspects of the S3 workflow including looking up keys for based
on an AWS instance's data. Here we move the AWS settings into a profile for
development and will retreive the same information via the instance in staging
and production.