Ensure user id has been prepended to api request.
We don't want any user to be able to upload files and we don't want malicious
users to be able to overwrite files from another user. We'll use uuids for the
filenames uploaded to S3 and prepend them the user's id number. We don't want
to rely on JS to tell us the truth so we'll verify here that the user is who
they say they are before we sign the request.