Conditional UI elements & split components.
We can reduce similar code by creating a few more components and moving the
secondary buttons into a reusable chunk. This greatly reduces the size of the
original component.

Some UI elements are only available if the logged in user is the owner of a
layer. This creates basic support for conditional UI elements within the react
component.