Trigger rerender on user state change.
The user is not logged in when the page first loads. We determine the current
user by querying an API endpoint asynchronously. This change allows the React
component to rerender when the logged in user changes and thus shows the UI
tools that are available when a user owns the layer.