Adjust secondary button functionality.
Secondary buttons are triggers to slide open UI elements for searching and for
sorting. This is handled through bootstrap but requires unique ids to be
assigned to the various elements. Since we reused large sections of code in
the components these ids needed to be generated dynamically to allow these
buttons to work properly.