Add views, routes, controllers, models for registration
Note that a new (relative to MMW) activation workflow was added. The activation link sent to users now leads to a custom back-end view which does the activation logic and then redirects to a front-end view which says that activation is complete. This was done in an attempt to render all HTML on the client-side for the sake of consistency. Not sure if it was worth the effort...

This also differs from MMW in that each screen has its own route so that the back button works, etc.

I attempted to factor out code common to the different screens using a Mixin, but due to the lack of inheritance in React.js, and the fact that mixin functions cannot be overridden, there is still some duplicated code. Hopefully we can find some design pattern to get around this.