Implement 'Select All' functionality.
The library UI has a checkbox at the top that acts as a toggle to select or
deselect all items in the library, eventually allowing the user to delete all
the items or do other operations on them. This wires the header checkbox to
toggle all checkboxes on child components.