Implement sorting functionality tied to UI.
The UI has a dropdown that allows for the list of layers in a tab to be sorted
based on the value in the dropdown. This adds some dummy data to assist with
testing and provides a mechanisms to sort the collection and rerender the
react components based on both the filters and the sorting mechanism chosen.