Remove user ping
This removes the user fetch request that occurs every time you switch
between views. The delayed fetch would sometimes cause a noticable
"flash" as the Library would load for a split second before being
replaced by the login screen (if not authenticated).

The user model is bootstrapped with data embedded in the HTML harness.