Fix background video reset on login screens
React will re-render the entire DOM node if you pass in a new kind of
React component, even if they generate nearly identical markup. This is
problematic because of the way our components were using composition to
share common markup.

The solution is to invert the common React components so that only the
inner content is re-rendered for each transition.

After refactoring out `UserScreen` from all these components I started
to receive the following error:

> Invariant Violation: ReactMount: Two valid but unequal nodes with the same `data-reactid`

I think this is triggered by components that return a different kind of
root node depending on the circumstances (ex. sometimes all content is
wrapped in `<div>` and other times `<form>`). I solved this by adding a
wrapper around each component output so the root node is consistent.

Connects #70