Fix screen transitions
Screen transitions were not being redrawn because the React render
function was being called inside a memoized function. The solution is to
split these apart so that the model/properties are memoized (we want
these to be persisted for the entire lifetime of the application) but
the React render calls are not. I also created a decorator called
`libraryView` to reduce some boilerplate.

Note that this temporarily breaks layer selections. This will be
addressed in the next commit.

Connects #110
Connects #113