Fix tile layer toggle
This restores the ability to add tile layers to the map by clicking on a
layer item in the sidebar.

This feature has been re-worked. Instead of adding a tile layer for each
checked layer in the sidebar, now, only 1 tile layer may be active at a time.
The tile layer will be added to the map by clicking on the layer item
instead of marking the checkbox.

The result is that clicking on a layer item will open the detail panel
and add a tile layer to the map. Closing the detail panel will remove
the tile layer from the map.