Make processor class poll SQS & route data.
This creates a function to kick off polling the SQS queue for messages. When a
message is received, it is directed to the proper processing function which in
turn will add another message to the queue if necessary. Messages that do not
conform to the proper data structure are deleted from the queue by default to
prevent them from piling up. The polling is handled in a simple loop. If no
messages are found in the queue, a waiting period is incremented by one second
and then a timeout is set. The max waiting period is one minute. Which is
a tolerable time to wait for a job to start but is not wasteful if there is
no work to do for a long time.

The structure of the SQS messages was changed to use only JSON. This was
decided because this is the structure used by S3 when posting to SQS so we
follow a similar JSON structure. This allows a single check to determine how
to route the data through the pipeline.