Fix favorited layers not appearing as favorited
Fixes a bug where favorited layers that were authored by other
users would not appear as favorited.

This inconsistency was caused by erroneous logic that was used to
populate the `favorite` property on layer models. After reflecting on
this issue, I realized that it was a mistake to add this propery to the
layer model. Instead, we now define a list of favorites on the user
model.

The best way to test this is by running the sample data generation
script. On develop, you should see that there are some layers on the
Favorites panel where the star is not illuminated. Switching to this PR
should correctly highlight all favorited layers on this panel,
regardless of the layer author.

Connects #150