Use contentType='application/json'
When posting layers, use contentType=‘application/json on the
front-end, backend, and unit tests. This is to more easily handle
nested data and not need the PatchedQueryDict workaround.

Also, make PUT work and add unit test. This fixes a bug where we tried
to access request.PUT, which is no longer needed since sending the data
as type ‘application/json’.