Worker task to validate that images have 3 bands.
As part of the worker job pipeline, we need to ensure that each tiff has three
bands. We use gdal to test for this. The first Mb of data has enough data to
verify the image bands so we get the first million bytes from the tiff on S3
and store it into a temporary file. This temp file is loaded with gdal and the
bands are counted. We add a new INVALID status to the layer image enum and
update the django model if the image cannot be loaded or three or more bands
are not detected.