Add dismiss endpoint and add live layer filter.
We don't want to return failed or processing layers in the layer selection
boxes since they can't be used. Adjust the endpoints to accept a filter for
"Live" (completed) layers and adjust the front end requests to use this new
filter.

Create an endpoint which will is used to "dismiss" jobs from the front end. If
the layer is failed, this deltes the layer. If the layer is completed
successfully, then this is a no-op. This is done to ensure that good layers
aren't accidentally deleted because of strange front end behavior.