Make front end error handling more versatile.
We had been hard coding the expected errors from the API. This is problematic
if we update the API to do better error handling. Now, we capture all errors
from the API and set them as appropriate on the front end. This is handled by
looping over the keys of the returned error object and setting them on our
front end model as needed. Errors still need to be added to the react
component to be properly displayed but no extra logic should be needed to get
the correct ones once they are returned.