Add monitoring endpoint.
To help us keep track of service interruptions we add a monitoring endpoint
that lets us know if the application stack and database are working properly.

We currently will only have a single app server, we return 200 OK for the
apps since be default of the fact that our endpoint code is running, the
application server must be ok.

We return a 200 for the database if we can connect to is. Currently we have
and expect to have only one database as well.

Worker servers are not monitored currently.

Updates cloud formation elb healthcheck endpoint config.