Use boto to check EMR cluster and step statuses.
Rather than trying to determine a reasonable timeout period, we use boto to
check on the status of the EMR cluster and the job steps in the cluster. If
the cluster or the steps seem to be in a non-running state, mark the job as
failed. We keep the timeout around because there are many unknowns and places
where a failure could result in hanging jobs.