Wrap boto connection and response in try/except.
There are many things that could go wrong with the boto connection and cluster
check. Key errors, bad responses, connection errors. Since any of these could
mean that the client failed, we wrap this chunk in try with generic exception
capturing and mark the job as failed if we fail for any reason.