Create status columns for layer model.
In order to keep better track of the status of jobs, we want to mark each
task as complete independently of the other ones. As part of this we also add
an associated error message to each status field. This is achived by adding a
start and end timestamp field for each status we use. As jobs move through the
pipeline they can update these timestamps to show that a job started or
completed.