Adjust naming conventions to be more accurate.
Concepts such as preValidated are no longer used. Now we have statuses for
uploading. Uploading also should not be part of the logic used to determine if
a layer is currently still processing. This cleans up some of the logic
surrounding classes in the processing block and adjusts the names to be more
in line with the current functionality.