Add retry endpoint
For failed layers, the retry endpoint resets status fields and starts
validation jobs for uploaded images, and copy jobs for s3 imports.

A new job was added for image arrival which triggers validation jobs
(which are now created by us, not S3). This is to make it easier to
manually add validation jobs.