Throw error if trying to fail a completed job.
If we are marking the layer as failed, we need to ensure that it has not
already been marked as completed. Raise an exception if the any code uses the
model instance methods to mark completed layer as failed.

As part of this, ensure that if a layer being marked as complete, it unsets
the failure status and failure error messages.