Add heartbeat status field to layer model.
We want to ensure that we only mark a job as failed if it received two
failing heartbeats in a row. This is necessary because in some cases, because
the queue is processed in order, we process a heartbeat job before the success
message has been processed in the queue, even though the queue has received
the message.