Two failed heartbeats needed to mark job failed.
Since a heartbeat queues the next heartbeat and since we sometimes receive a
message that the job has completed in an order such that our heartbeat
incorrectly thinks the job has failed, we will only mark the job as failed if
two heartbeats return a failed status. This way if a completed message returns
in between heartbeats, the job will not be incorrctly marked as failed.