Expand processing block on adding layer
After adding a layer, the processing block should expand so the user
knows that it is possible to expand it, and to show that the layer is
being processed.

This implementation is not in the spirit of React, but I couldn’t
figure out how to preserve the animation and styling that comes with
Bootstrap in a React-style implementation.