Add bounding box info to image & layer models.
We need to know the maximum and minimum bounds for each image and for the
layer as a whole so that we can zoom to the layer bounds on the front end.
This adds support for the bounding box as four new fields on both the image
and layer models.