Find and save bounding box and other metadata.
When processing images there are two keys points to save different kinds of
metadata. The first is during image validation when we have already retrieved
the tiff headers via gdal and vsicurl. This fast and efficient since gdal only
gets what is needed to get the info requested. While we have gdal verifying
the image we can get the bounds from it and store them on the image model.

The other key point to get metadata is during thumbnailing. During this phase
we have the entire image on disc already. Before it is erased, we can process
the exif info from the tif.

Finally, when all the images have been verified, we know that the full extent
of the layer is known so we calculate the max and min bounds for all the
images to calculate the extent of the entire layer.