Fix IOError during thumbnail phase
This fixes a bug where certain TIFFs were not able to be thumbnailed by
Pillow. I solved this by using `gdal_translate` to produce a PNG which
is then used to generate each thumbnail.

I also discovered a utility function called `fit` in the PIL library
which I have decided to add since it is more succinct and also gives us
the ability to center align thumbnail images.

Connects #333
Connects #340