Fix thumbnail invalid UUID exception
Because thumbnail filenames now contain the UUID of the original image
(instead of a random UUID) the queue processer was attempting to
revalidate the original image every time a new thumbnail was uploaded to
S3.

Now, we assume that if LayerImage does not exist in the DB (for whatever
reason) it must be a thumbnail.