Add initial swagger specification for user/orgs
This commit adds an initial swagger specification for users and
organization endpoints.

The user model is kept minimal for now, assuming relevant information is
going to be stored in Auth0 and returned (name, email, etc.); however,
the fields we will want to use for foreign key constraints are added
here (organization, role).

Additionally, a docker container is added for the swagger editor to ease
development that will be available in the browser. This is started via
docker-compose and our scripts.