Add initial Auth0 settings to application
This commit adds support for using Auth0 settings in the application:
 - client id: used to authenticate against a certain application in
   Auth0
 - domain: domain to use for sending requests to Auth0 for
   authenticatino
 - secret: secret used to verify signature of JWT

In this case the client secret is only the secret for the development
environment. Because this accesses an external service it still needs to
be kept secret. To accomplish this environment settings/files are placed
in an S3 bucket that is pulled down during provisioning. This requires
that developers have access to the AWS account.