Use SWF for async tasks/workflows
This commit adds support for using SWF to run asynchronous tasks and
workflows using the botoflow library. This commit includes stubbed out
functionality for handling and kicking off what could be basic tasks for
processing image uploads.

The `click` library is used for building a command line client for
starting activity and decider workers that process jobs as they come
in.

An SWF domain is created for developers for development and a single
activity and decider worker is started with docker compose.