Add initial airflow scaffolding
This commit adds a prototype for using airflow for task management. A
simple DAG is added to simulate an import of a single geotiff that would
cause 3 tasks to be kicked off:
 - metadata extraction
 - footprint extraction
 - thumbnail generation

Airflow itself has a web UI that this commit adds for opening up on
port 8080. Flower is used for monitoring celery workers, which the
airflow worker uses for parallelization of tasks. A redis instance is
added as a broker for the celery workers.