Add authentication directive to authenticate users
This commit adds two directives to handle authentication.
 1. Authenticates if credentials are provided, but allows anonymous
    users
 2. Requires credentials

The first method is useful for when we want to use information about a
user if it is available, but not require it and the second is if we
require information.

For method 1 if a JSON web token provided in the Authorization header then one of
three things can happen:
 1. Token Valid and User in Database: request proceeds
 2. Token Valid and User _not_ in Database: rejected
 3. Token Invalid: rejected

If no Authentication is provided then an anonymous user is attached to
the request.

For method 2 everything about method 1 still applies; however, if no
authentication is provided then authentication fails.

Also slightly cleans up the routes by converting them to traits.