Add pagination to users endpoint
This commit adds initial support for pagination for the users API
endpoint.

It adds a new case class for paginated responses that is parameterized
on type so that it can be used by any API endpoint that needs
pagination.

Additionally, adds support for sorting on organization, email, and  user
id. This is not likely to be used in our case; however, I think it
illustrates an approach enough to make it easy to extend/adapt for
additional fields we want to allow sorting on for other API endpoints
eventually.