Removed unset list item by default in docker task
This fixes an issue where the step to add users to the docker user group
was not running because the `jenkins_name` variable was unset in a
vagrant environment. This change fixes it by defining the set of users
to add to the group via group variables that are specific to each
environment.