Add functions/module to extract landsat footprint
This commit includes all the functionality required to extract
footprints from landsat8 data. It accomplishes via the following
workflow:
 - downloads a landsat scene from s3
 - creates a tif mask of where 0 is nodata and 1 is data
 - extracts polygon from tif mask

Functions to get and post information to raster foundry via API
endpoints are roughly stubbed out, but will need to be adjusted once the
exact workflow of landsat 8 ingestion is decided.