Fix CI test execution
Build Airflow container image, and then execute test suite. Also, ensure
that the Git SHA is threaded through to `test` so that the correct
container image is used.

Lastly, create a volume mount for the `app-tasks/rf` so that `setup.py`
exists within the container and can be used for testing.