Add buckets, refine layers/images API endpoints
The big change here is the addition of the concept of scenes and removal
of imports.

Scene in this case is a group of images that are related in some way,
sharing large portions of metadata. There are two examples that
illustrate the need for this type of resource:

 - pre-tiled imagery: in this case a user uploads imagery that is split
   for filesize purposes, but not for any other reason. This imagery
   shares the same metadata, but users would expect to interact with all
   the images together and we would like to ingest it via GeoTrellis
   together as well. From a UX perspective, the advantage here is that
   if 100 images are uploaded with the same metadata and user does a
   search where the primary unit is an "image", all 100 images would be
   returned. However, if grouped as a scene, only one result would be
   returned.
 - landsat 8 imagery: in this case, one image per band is uploaded, but
   logically they are all part of the same scene and share metadata

Other changes include updating query parameters for users since it is no
longer possible to query users by organization or role when they have
multiples of both. To get users that belong to an organization requires
querying the organization.