Add Scenes model and helper trait
This commit adds a 'Scene' model that represents a group of images with
shared metadata that has thumbnails, a boundary, and shared metadata.

Scenes are the primary interaction for users. See the swagger spec for
more information about the meaning of fields.

Also moves generated_code => datamodel to be less confusing  since inevitably
some files will need to be edited here.