Add buckets, buckets to scenes relations
This commit adds buckets and a relationship between buckets and
scenes. Buckets are an arbitrary grouping of scenes that users may want
to perform some action on or save for later references.

Basic database interaction is added as well with the ability to
create/read/update/delete buckets in addition to adding, listing, and
removing scenes from a bucket.