Add name attribute to scene
While not all that much clearer than a random UUID, identifiers like
LCXXXXXXXX for landsat 8 and similar identifiers for other imagery can
be useful in identifying or associating scenes at a glance. This commit
adds a name attribute with the intention that these identifiers can be
used there -- with the added benefit of being able to enforce some
uniqueness constraints that will help prevent importing the same data
multiple times.