Split out related objects into Models.scala
With all of the joins and the ability to create related objects with a
single post there was a proliferation of case classes. These area all
moved to Model.scala to keep things a little clean; however, I expect
some of these to get cleaned up in the next refactor as we stabilize the
data access model.