Add support for related models with scenes
This commit does two big changes to the API:

 1. Adds support to include thumbnails, footprint, and images when
 creating a scene. This simplifies things somewhat because it allows
 users to create a scene without having to make many POST requests to a
 number of different endpoints. However, these are optional such that if
 a user desires to construct things in a piecemeal fashion they can
 still do so.

 2. Adds support for returning footprints, thumbnails, and images in the
 response for a scene. This will simplify frontend and client API
 interaction by providing all of the necessary information in a single
 API call.

After adding thumbnails, footprint, and images in the scene result
pagination was no longer working because of the server side reduce.
This commit fixes that problem by doing the filtering, pagination, then
an inner join on the related models so that when we do the server side
reduce/group by the number of results we want will always be the page
size that was requested.