Return 400, not 500 for invalid tokens
Previously we would return a 500 for every Auth0 API error. But
there is a use case when the user sends an invalid refresh token
and Auth0 returns a 401 with an invalid_refresh_token error. Since
this is a legitimate user error, we now inspect the Auth0 response
and send back a 400 if this is the case.

We also add the UserErrorHandler trait and respective exception
handler to the token route so it can translate an
IllegalArgumentException to an error 400. Also update the logging
code so now it prints the full stack trace for every error.