Use Band model, resolution, and metadata file on import
Models for the importer had become stale relative to models in the scala
backend, so imports didn't work for a variety of reasons. This commit
modifies the whole import process from `models/*.py` to the upload
scripts and modified the scala backend to deal with the updated incoming
data.

Summary of changes:

- synced fields between `x.scala` in the datamodel and `x.py` in
  importer
- updated scene insertion logic to deal with presence of images with
  related
- updated the style guide to include style for "messy" joins
- included geotrellis packages in `app-backend/app/build.sbt`