Add ability to filter scenes by image params
Now in addition to scene specific parameters, the scene endpoint
can also take image parameters. This will result in selecting only
those scenes that have images which satisfy the filter. The images
themselves will not be filtered.

For example, consider a scene S with images I1 and I2, whose
resolutions are 10m and 20m respectively. If a minResolution=15
filter is applied, S will match because I2 satisfies the
constraint. When S is returned in a Scenes.WithRelated format, it
will include both I1 and I2, even though only I2 satisfied the
filter. We filter scenes, not images.