Update swagger spec to accomodate API changes
A few API endpoints have been edited since the swagger spec was
written. The goal of this commit is to get the spec more in line with
how the API has been implemented and also eliminated some
inconsistencies.

Summary of changes:
 - Added Authentication tag and removed Organization
 - Removed boundaries model
 - Changed references to boundary to scene
 - Added band model
 - Added endpoints for tokens
 - Updated buckets endpoints to reflect bulk API implementation

Additionally, new endpoints need to be sketched out to support the model
catalog

This includes the following:
 - model endpoint for geoprocessing
 - model tags endpoint to store tags
 - model category endpoints to store categories

Further, the following query parameters in anticipation of supporting the
wireframes for the model catalog.
 - modelTag -- allows to search by tags
 - modelCategory -- allows to search by category
 - minRating -- returns only models with a rating greater than a certain
 number
 - maxRating -- returns models only less than a certain number
 - search -- full-text search query parameter

Omitting price because we are only treating this as a catalog and not a
marketplace in the short term