Automatically filter scenes based on map view box
The LeafletMap component now publishes an onBoundsChange binding to
which the parent BrowseController listens for updates. This is then used
to update the query params and filter the scene list by the bounding box
of the map view. If valid bounding box coordinates are passed in as
query parameters on initial page load, they will be used to set the
initial map view.