Add model categories for model lab models
This commit adds a migration and related scala code to support inserting
and retrieving a model category from the database.

Note: Because model categories are more strict than model tags and can
be only created by Raster Foundry, not users, the organization ID
parameter is ommitted with the understanding that we will not allow
restricting categories to an organization or imply ownership.