Ensure that Airflow doesn't run as root
Create an `airflow` user that is responsible for running the various Airflow
daemons. Ensure that its GID is 1000 so that it can read the AWS credentials in
the development environment. Also, change the volume definition for the AWS
credentials so that they map to the `airflow` user's home directory.

Lastly, tighten constraints around Celery installation so that version 4.x isn't
installed.