Filter scenes by `createdAt` to prevent inconsistent paging
This change caches a timestamp each time `populateInitialSceneList` is called.
This timestamp is used as the value for the `maxCreateDatetime` parameter within
the page query. This ensures that any scenes created after the first page is returned
will not offset the query results and cause duplication in the list when subsequent pages
are requested.