Hook color correction UI to tileserver
This commit hooks up the color correction UI to edit both a single and
to bulk edit scenes for color correction.

This does not hook up the histogram, an issue has been added for that.

For the time being, the false color composites can be used for ingested
Landsat 8 scenes, but not other ingested layers. This likely needs to be
dynamic and determined by datasources.

The reset button is still broken slightly - the layer resets;
however, the sliders are not reset. The current set up makes this
difficult because of how the static sidebar is set up. Rather than
creating more issues for the next task (fix routing/back button) I added
a todo.