Allow selecting scenes to color correct by grid
Displays a highlight on the map to show the selected grid cell and
allows toggling the selected area by clicking on it.

Also moves the code for retrieving all of a Project's Scenes into the
ProjectService and refactors the mosaicking and color correction
interfaces to use the ProjectService method.

Fixes an incorrect query parameter being used in the
ProjectService.getProjectSceneCount() method that was causing it to
return a full page of results; it now returns only a single result and
is much faster.