Toggle scene selection when a scene is clicked on
What should happen when a user clicked on an unselected scene from a
list after having selected some scenes was difficult to decide, and
there was no indication that clicking on different parts of scenes in
the list in the color correction view would lead to different outcomes.
This commit makes it so there aren't different outcomes.