Fix production build styling
This commit fixes the production build styling by switching to the
extract text plugin for styles during production builds. There may be a
small performance penalty, but this solves the styling issue by ensuring
that all the styling is concatenated together into a single file. The
project's styles are applied last (the next sass style loader) so our
overrides will be preserved.