Replace scene grid endpoint
This commit replaces the current scene grid endpoint/implementation with
one that uses Leaflet's grid layer to request summary aggregate
information for each tile on the map. This speeds up the endpoint by
reducing the amount of information required to power it.

The grid layer itself uses Konva to draw rectangles and style
them. Canvas events are used to display the count when the canvas
rectangle is hovered over. Additional styling work is necessary, but
this gives us hopefully a start.