Retrieve/save color corrections from API
This commit adds support for retrieving and saving color corrections
from the API. A service was added to interact with the API, but will
need to be refactored once #880 is completed and the API is fixed to be
slightly more idiomatic.

Every change to a slider is synced with the API so that users do not
need to worry about "saving" their work.