Replace union with or filter for user visibility
This commit replaces the previous implementation of using a union of two
select statements with a filter that uses an OR condition to filter
records according to if the record is labeled as PUBLIC or is created by
the user making the API request.

A new trait is added that combines the UserFK trait with Visibility that
can be used for tables that have both a visibility field and foreign key
for users.