Add image overlay clipping to thumbnail display
This commit adds a custom image overlay layer for the map that uses
canvas to clip approximately to the shape of data for a thumbnail. This
allows Raster Foundry to continue using thumbnails generated from
Landsat 8 and Sentinel 2 but display them in a more realistic/true
manner.

Unfortunately, it seems really difficult to eliminate all of the nodata
black/white from the thumbnails without a significant chunk of work.