Use correct mosaic tile URL in publish modal
This commit adds support for inserting the correct tile URL in the
publish modal.

There is a problem with the mosaic tile server such that it requires the
user id of the user with scenes; if anything, this needs to be the user
of the project, not the individual scenes since a mosaic may include
tiles from many different users.