Persist scene footprints in color correct
Provides the user with feedback regarding which scenes they've selected
by persisting the selected scenes' footprint to the map. This should
make it clearer why the map decides to zoom the way it does. This trades
off some efficiency in the selectAllScenes / selectNoScenes methods in order to
keep the code simple, but it's still fairly snappy even when selecting
all of the Egypt scenes. If it becomes a performance bottleneck we may
need to adopt something a bit more complex.