Add `ingestLocation` to scenes, allow filtering
This commit also drops the `status` field for scenes.
It adds a query parameter, `ingested`. When ingested is set to 'true',
only scenes that have been ingested (have a non-null `ingestLocation`)
will be returned. When ingested is set to 'false', only scenes that have
not been ingested (have a null `ingestLocation`) will be returned.