Add "proof" of concept actions refactor
This was supposed to remove a lot of complexity in our database handling, but
more than anything else it seems to have shifted that complexity elsewhere
without paying off in reusability as I'd hoped it would. Because of our complex
types (not complex in logic, but complex in the sense of "T.WithRelated"), there
isn't a way to go to the database and return a DBIO for a lot of things we might
want. For example, you can't go to the database and get an Image.WithRelated in
a DBIO to combo with a Scene. This wasn't a _waste_ of time, but whatever
unpleasantness we think our current design patterns bakes into developing
against the database is here to stay for a moment at least.

There is a useful pattern that came out of this at least, and that's the Aux
pattern. The aux pattern lets us define type viewability in terms of what types
can do instead of what they are, as long as there's an implicit value meeting
the requirements of the typeclass in scope. Where we have common/repetitive
functionality, we could probably abstract it into a typeclass.