Create and alter permissions on AIRFLOW_HOME
Previously, the `AIRFLOW_HOME` directory was being created by the `adduser`
command, which created it and ensured that the `airflow` user was given correct
permissions.

Because of the new disconnect between `airflow` user `HOME` and `AIRFLOW_HOME`,
the `AIRFLOW_HOME` directory needs to be created (before `COPY`) and the
`airflow` user granted proper permissions.