Fix auth0 management tokens
This commit alters the approach to obtaining a bearer token that granted access to the auth0 management api. We utlize the management api to create device_tokens which we leveraged when created user's api tokens.

Auth0 moved to a more secure token implementation which adopts a 24-hour lifetime token rather than the long-lived token used previously. As a result, we can no longer set an environment variable with the token and use that when doing device_token CRUD operations.

This commit uses an async-cache with a 12-hour expiration time (adjustable) to handle getting and storing the current management bearer token.