Refactor color correct API for updates
Color corrections are set to defaults/empty objects when a scene is
added to a project. Additional endpoints are provided as a convenience
to update only color corrections or mosaic order. Before this commit,
those updates were handled through a POST request that returned an empty
object. This PR refactors the endpoint and frontend to use PUTs in a
more idiomatic manner.

A few changes to the response JSON were made to make it easier for
clients to parse responses and post initially as well.

fix tests