Remove call to get JWT in airflow scheduler
This commit removes the call to get a JWT from the "find scenes to
ingest" DAG. This caused an error in the airflow scheduler because the
airflow scheduler does not have the Auth0 client token injected into its
environment. Instead, this commit moves the creation of a session inside
a function to ensure it only gets called when run inside an Airflow worker.