Add new create project modal flow
This PR adds a multi-stage project creation modal that is initiated from the home page when the user
clicks the 'Create a new Project' button. The import step of the process is stubbed, and has no effect
on the project creation process. If the user chooses the 'Public Imagery' option, they are taken to the
browse page.