Clean up dangling underscore styling
Our practice is to use dangling underscores to denote private variables,
but this wasn't reflected in our ESLint configuration, which led to lots
and lots of comments to disable the warning scattered throughout our
code. This changes the global configuration to allow dangling
underscores only after `this`, and then adds a few overrides that are
necessary for dealing with Leaflet.