Add ability to read direct values to ActionRunner
This is useful when a query always returns a NOT NULL value,
such as when running a COUNT. Basically, instead of taking
Option[T] like readOne, it takes T itself.

Notably, the type signature of this method is identical to
`write`, and thus I could just have used `write`, but that
would not have been an accurate representation of intended
activity, thus I decided to create a new one.

I also tried to overload `readOne`, but since Option[T] is
just an instance of T, the newer definition would subsume
the other one.